refactor: update progress reporting function signatures to use int64_t for processed values

+++ src/cpp/dfs.cpp
@@ -58,7 +58,7 @@ JsBestMixResult findBestMixDFSJsonWithProgress(
   applySubstanceRulesJson(substances, substanceRulesJson);
 
   // Function pointer for progress reporting in console mode
-  extern void reportProgressToConsole(int depth, int processed, int64_t total);
+  extern void reportProgressToConsole(int depth, int64_t processed, int64_t total);
 
   // Run the DFS algorithm with progress reporting if enabled
 #ifdef __EMSCRIPTEN__
